services/api-gateway/pyproject.toml
[tool.poetry]
name = "hopper-api-gateway"
version = "0.1.0"
description = "Hopper GPU Cloud Platform - API Gateway"
authors = ["Hopper Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
pydantic = "^2.9.0"
pydantic-settings = "^2.5.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.0"}
asyncpg = "^0.30.0"
alembic = "^1.14.0"
nats-py = "^2.9.0"
grpcio = "^1.65.0"
grpcio-tools = "^1.65.0"
structlog = "^24.0.0"
slowapi = "^0.1.9"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
httpx = "^0.27.0"
sse-starlette = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.24.0"
pytest-cov = "^6.0.0"
ruff = "^0.8.0"
testcontainers = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
